+++ ElvUI_MerathilisUI/Core/Changelog/6.47.lua
@@ -1,8 +1,10 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[647] = {
-	RELEASE_DATE = "TBD",
+	RELEASE_DATE = "19.12.2024",
 	FIXES = {},
 	NEW = {},
-	IMPROVEMENTS = {},
+	IMPROVEMENTS = {
+		"[Misc]: Bump toc",
+	},
 }