+++ ElvUI_MerathilisUI/Core/Changelog/6.53.lua
@@ -2,7 +2,12 @@ local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[653] = {
 	RELEASE_DATE = "TBD",
-	FIXES = {},
+	FIXES = {
+		"[11.1]: Changes because of the new Patch",
+		"[World Map]: Disable Event Tracker for now, Blizzard got their own one",
+	},
 	NEW = {},
-	IMPROVEMENTS = {},
+	IMPROVEMENTS = {
+		"[AutoButtons]: Add more openable items",
+	},
 }