+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 ### Changes ###
 
-  * [KeyFeedback]: Should now work again
-  * [ObjectiveTracker]: Wowheadlinks should now redirect to the english comments
-  * [Skins]: Updated various skins
+  * [Misc]: Bump toc
+