+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Changes ###
 
-  * [Objective Tracker]: Fix typo in the options
-  * [Wowhead Links]: Now also works with monthly activities
+  * [11.1]: Changes because of new Patch
+  * [Wowhead Map]: Disable Event Tracker for now, Blizzard got their own one
+  * [AutoButtons]: Add more openable items